Enhancement: Turn Sniffer into named constructor on Format value object
