--- src/Format/SnifferInterface.php
@@ -1,21 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * Copyright (c) 2018 Andreas Möller.
- *
- * For the full copyright and license information, please view
- * the LICENSE file that was distributed with this source code.
- *
- * @see https://github.com/localheinz/json-normalizer
- */
-
-namespace Localheinz\Json\Normalizer\Format;
-
-use Localheinz\Json\Normalizer\JsonInterface;
-
-interface SnifferInterface
-{
-    public function sniff(JsonInterface $json): FormatInterface;
-}
