+++ test/Unit/Format/FormatTest.php
@@ -17,6 +17,7 @@ use Localheinz\Json\Normalizer\Format\Format;
 use Localheinz\Json\Normalizer\Format\FormatInterface;
 use Localheinz\Json\Normalizer\Format\IndentInterface;
 use Localheinz\Json\Normalizer\Format\NewLineInterface;
+use Localheinz\Json\Normalizer\JsonInterface;
 use Localheinz\Test\Util\Helper;
 use PHPUnit\Framework;
 
@@ -184,6 +185,402 @@ final class FormatTest extends Framework\TestCase
             yield $key => [
                 $hasFinalNewLine,
             ];
+        }
+    }
+
+    /**
+     * @dataProvider providerJsonEncodeOptionsAndEncoded
+     *
+     * @param int    $jsonEncodeOptions
+     * @param string $encoded
+     */
+    public function testFromJsonReturnsFormatWithJsonEncodeOptions(int $jsonEncodeOptions, string $encoded): void
+    {
+        $json = $this->prophesize(JsonInterface::class);
+
+        $json
+            ->encoded()
+            ->shouldBeCalled()
+            ->willReturn($encoded);
+
+        $format = Format::fromJson($json->reveal());
+
+        $this->assertInstanceOf(FormatInterface::class, $format);
+        $this->assertSame($jsonEncodeOptions, $format->jsonEncodeOptions());
+    }
+
+    public function providerJsonEncodeOptionsAndEncoded(): array
+    {
+        return [
+            [
+                0,
+                '{
+  "name": "Andreas M\u00f6ller",
+  "url": "https:\/\/github.com\/localheinz\/json-normalizer"
+}',
+            ],
+            [
+                \JSON_UNESCAPED_SLASHES,
+                '{
+  "name": "Andreas M\u00f6ller",
+  "url": "https://github.com/localheinz/json-normalizer"
+}',
+            ],
+            [
+                \JSON_UNESCAPED_UNICODE,
+                '{
+  "name": "Andreas Möller",
+  "url": "https:\/\/github.com\/localheinz\/json-normalizer"
+}',
+            ],
+            [
+                \JSON_UNESCAPED_SLASHES | \JSON_UNESCAPED_UNICODE,
+                '{
+  "name": "Andreas Möller",
+  "url": "https://github.com/localheinz/json-normalizer"
+}',
+            ],
+        ];
+    }
+
+    /**
+     * @dataProvider providerEncodedWithoutIndent
+     *
+     * @param string $encoded
+     */
+    public function testFromJsonReturnsFormatWithDefaultIndentIfJsonIsWithoutIndent(string $encoded): void
+    {
+        $json = $this->prophesize(JsonInterface::class);
+
+        $json
+            ->encoded()
+            ->shouldBeCalled()
+            ->willReturn($encoded);
+
+        $format = Format::fromJson($json->reveal());
+
+        $this->assertInstanceOf(FormatInterface::class, $format);
+        $this->assertSame('    ', $format->indent()->__toString());
+    }
+
+    public function providerEncodedWithoutIndent(): \Generator
+    {
+        $values = [
+            'array-empty' => '[]',
+            'array-without-indent' => '["foo","bar baz"]',
+            'bool-false' => 'false',
+            'bool-true' => 'true',
+            'float' => '3.14',
+            'int' => '9000',
+            'null' => 'null',
+            'object-empty' => '{}',
+            'object-without-indent' => '{"foo":"bar baz","baz":[9000,123]}',
+            'string-blank' => '" "',
+            'string-word' => '"foo"',
+        ];
+
+        foreach ($values as $key => $value) {
+            yield $key => [
+                $value,
+            ];
+        }
+    }
+
+    /**
+     * @dataProvider providerPureIndentAndSniffedIndent
+     * @dataProvider providerMixedIndentAndSniffedIndent
+     *
+     * @param string $indent
+     * @param string $sniffedIndent
+     */
+    public function testFromJsonReturnsFormatWithIndentSniffedFromArray(string $indent, string $sniffedIndent): void
+    {
+        $encoded = <<<JSON
+[
+"foo",
+${indent}"bar",
+    {
+        "qux": "quux"
+    }
+]
+JSON;
+
+        $json = $this->prophesize(JsonInterface::class);
+
+        $json
+            ->encoded()
+            ->shouldBeCalled()
+            ->willReturn($encoded);
+
+        $format = Format::fromJson($json->reveal());
+
+        $this->assertInstanceOf(FormatInterface::class, $format);
+        $this->assertSame($sniffedIndent, $format->indent()->__toString());
+    }
+
+    /**
+     * @dataProvider providerPureIndentAndSniffedIndent
+     * @dataProvider providerMixedIndentAndSniffedIndent
+     *
+     * @param string $indent
+     * @param string $sniffedIndent
+     */
+    public function testFromJsonReturnsFormatWithIndentSniffedFromObject(string $indent, string $sniffedIndent): void
+    {
+        $encoded = <<<JSON
+{
+"foo": 9000,
+${indent}"bar": 123,
+    "baz": {
+        "qux": "quux"
+    }
+}
+JSON;
+
+        $json = $this->prophesize(JsonInterface::class);
+
+        $json
+            ->encoded()
+            ->shouldBeCalled()
+            ->willReturn($encoded);
+
+        $format = Format::fromJson($json->reveal());
+
+        $this->assertInstanceOf(FormatInterface::class, $format);
+        $this->assertSame($sniffedIndent, $format->indent()->__toString());
+    }
+
+    public function providerPureIndentAndSniffedIndent(): \Generator
+    {
+        $characters = [
+            'space' => ' ',
+            'tab' => "\t",
+        ];
+
+        $sizes = [1, 3];
+
+        foreach ($characters as $style => $character) {
+            foreach ($sizes as $size) {
+                $key = \sprintf(
+                    '%s-%d',
+                    $style,
+                    $size
+                );
+
+                $pureIndent = \str_repeat(
+                    $character,
+                    $size
+                );
+
+                yield $key => [
+                    $pureIndent,
+                    $pureIndent,
+                ];
+            }
+        }
+    }
+
+    public function providerMixedIndentAndSniffedIndent(): \Generator
+    {
+        $mixedIndents = [
+            'space-and-tab' => [
+                " \t",
+                ' ',
+            ],
+            'tab-and-space' => [
+                "\t ",
+                "\t",
+            ],
+        ];
+
+        foreach ($mixedIndents as $key => [$mixedIndent, $sniffedIndent]) {
+            yield $key => [
+                $mixedIndent,
+                $sniffedIndent,
+            ];
+        }
+    }
+
+    /**
+     * @dataProvider providerEncodedWithoutIndent
+     *
+     * @param string $encoded
+     */
+    public function testFromFormatReturnsFormatWithDefaultNewLineIfUnableToSniff(string $encoded): void
+    {
+        $json = $this->prophesize(JsonInterface::class);
+
+        $json
+            ->encoded()
+            ->shouldBeCalled()
+            ->willReturn($encoded);
+
+        $format = Format::fromJson($json->reveal());
+
+        $this->assertInstanceOf(FormatInterface::class, $format);
+        $this->assertSame(\PHP_EOL, $format->newLine()->__toString());
+    }
+
+    /**
+     * @dataProvider providerNewLine
+     *
+     * @param string $newLine
+     */
+    public function testFromFormatReturnsFormatWithNewLineSniffedFromArray(string $newLine): void
+    {
+        $encoded = <<<JSON
+["foo",${newLine}"bar"]
+JSON;
+
+        $json = $this->prophesize(JsonInterface::class);
+
+        $json
+            ->encoded()
+            ->shouldBeCalled()
+            ->willReturn($encoded);
+
+        $format = Format::fromJson($json->reveal());
+
+        $this->assertInstanceOf(FormatInterface::class, $format);
+        $this->assertSame($newLine, $format->newLine()->__toString());
+    }
+
+    /**
+     * @dataProvider providerNewLine
+     *
+     * @param string $newLine
+     */
+    public function testFromFormatReturnsFormatWithNewLineNewLineSniffedFromObject(string $newLine): void
+    {
+        $encoded = <<<JSON
+{"foo": 9000,${newLine}"bar": 123}
+JSON;
+
+        $json = $this->prophesize(JsonInterface::class);
+
+        $json
+            ->encoded()
+            ->shouldBeCalled()
+            ->willReturn($encoded);
+
+        $format = Format::fromJson($json->reveal());
+
+        $this->assertInstanceOf(FormatInterface::class, $format);
+        $this->assertSame($newLine, $format->newLine()->__toString());
+    }
+
+    public function providerNewLine(): \Generator
+    {
+        $values = [
+            "\r\n",
+            "\n",
+            "\r",
+        ];
+
+        foreach ($values as $newLine) {
+            yield [
+                $newLine,
+            ];
+        }
+    }
+
+    /**
+     * @dataProvider providerWhitespaceWithoutNewLine
+     *
+     * @param string $actualWhitespace
+     */
+    public function testFromFormatReturnsFormatWithoutFinalNewLineIfThereIsNoFinalNewLine(string $actualWhitespace): void
+    {
+        $encoded = <<<'JSON'
+{
+    "foo": 9000,
+    "bar": 123,
+    "baz": {
+        "qux": "quux"
+    }
+}
+JSON;
+        $encoded .= $actualWhitespace;
+
+        $json = $this->prophesize(JsonInterface::class);
+
+        $json
+            ->encoded()
+            ->shouldBeCalled()
+            ->willReturn($encoded);
+
+        $format = Format::fromJson($json->reveal());
+
+        $this->assertInstanceOf(FormatInterface::class, $format);
+        $this->assertFalse($format->hasFinalNewLine());
+    }
+
+    public function providerWhitespaceWithoutNewLine(): \Generator
+    {
+        $characters = [
+            ' ',
+            "\t",
+        ];
+
+        foreach ($characters as $one) {
+            foreach ($characters as $two) {
+                $whitespace = $one . $two;
+
+                yield [
+                    $whitespace,
+                ];
+            }
+        }
+    }
+
+    /**
+     * @dataProvider providerWhitespaceWithNewLine
+     *
+     * @param string $actualWhitespace
+     */
+    public function testFromFormatReturnsFormatWithFinalNewLineIfThereIsAtLeastOneFinalNewLine(string $actualWhitespace): void
+    {
+        $encoded = <<<'JSON'
+{
+    "foo": 9000,
+    "bar": 123,
+    "baz": {
+        "qux": "quux"
+    }
+}
+JSON;
+        $encoded .= $actualWhitespace;
+
+        $json = $this->prophesize(JsonInterface::class);
+
+        $json
+            ->encoded()
+            ->shouldBeCalled()
+            ->willReturn($encoded);
+
+        $format = Format::fromJson($json->reveal());
+
+        $this->assertInstanceOf(FormatInterface::class, $format);
+        $this->assertTrue($format->hasFinalNewLine());
+    }
+
+    public function providerWhitespaceWithNewLine(): \Generator
+    {
+        $characters = [
+            '',
+            ' ',
+            "\t",
+            \PHP_EOL,
+        ];
+
+        foreach ($characters as $before) {
+            foreach ($characters as $after) {
+                $whitespace = $before . \PHP_EOL . $after;
+
+                yield [
+                    $whitespace,
+                ];
+            }
         }
     }
 }
